不要なAPIレスポンス型を削除し、型定義をtypesに整理
