+++ src/types/record.ts
@@ -1,3 +1,4 @@
+// records/[date]/route.tsで使用中型
 export type RecordData = {
   id: number
   date: string
@@ -17,4 +18,18 @@ export type UpdateRecordRequestBody = {
   weight?: number | null;
   steps?: number | null;
   memo?: string | null;
+};
+
+// records/route.tsで使用中型
+export type CreateRecordRequestBody = {
+  date: string;
+  weight: number | null;
+  steps: number | null;
+  memo: string | null;
+};
+
+export type RecordsResponse = {
+  status: "OK" | "NG";
+  message: string;
+  records: RecordData[];
 };
